+++ firmware/prototyper/src/sbi/mod.rs
@@ -4,6 +4,7 @@ use spin::Once;
 pub mod console;
 pub mod hsm;
 pub mod ipi;
+pub mod mpxy;
 pub mod pmu;
 pub mod reset;
 pub mod rfence;
@@ -20,6 +21,7 @@ pub mod trap_stack;
 use console::SbiConsole;
 use hsm::SbiHsm;
 use ipi::SbiIpi;
+use mpxy::SbiMpxy;
 use pmu::SbiPmu;
 use reset::SbiReset;
 use rfence::SbiRFence;
@@ -42,6 +44,8 @@ pub struct SbiDispatcher {
     pmu: Option<SbiPmu>,
     #[rustsbi(susp)]
     susp: Option<SbiSuspend>,
+    #[rustsbi(mpxy)]
+    mpxy: Option<SbiMpxy>,
 }
 
 impl SbiDispatcher {
@@ -55,6 +59,7 @@ impl SbiDispatcher {
         rfence: Option<SbiRFence>,
         susp: Option<SbiSuspend>,
         pmu: Option<SbiPmu>,
+        mpxy: Option<SbiMpxy>,
     ) -> Self {
         SbiDispatcher {
             console,
@@ -64,6 +69,7 @@ impl SbiDispatcher {
             rfence,
             pmu,
             susp,
+            mpxy,
         }
     }
 }