+++ library/macros/src/lib.rs
@@ -29,6 +29,7 @@ struct StaticImpl {
     cppc: Option<Member>,
     nacl: Option<Member>,
     sta: Option<Member>,
+    mpxy: Option<Member>,
     env_info: Option<Member>,
 }
 
@@ -50,6 +51,7 @@ impl StaticImpl {
             "cppc" => (true, self.cppc.replace(member)),
             "nacl" => (true, self.nacl.replace(member)),
             "sta" => (true, self.sta.replace(member)),
+            "mpxy" => (true, self.mpxy.replace(member)),
             "info" | "env_info" => (true, self.env_info.replace(member)),
             _ => (false, None),
         }
@@ -69,6 +71,7 @@ struct DynamicImpl {
     cppc: Vec<Member>,
     nacl: Vec<Member>,
     sta: Vec<Member>,
+    mpxy: Vec<Member>,
     env_info: Option<Member>,
 }
 
@@ -86,6 +89,7 @@ impl DynamicImpl {
             "cppc" => self.cppc.push(member),
             "nacl" => self.nacl.push(member),
             "sta" => self.sta.push(member),
+            "mpxy" => self.mpxy.push(member),
             "info" | "env_info" => return self.env_info.replace(member).is_none(),
             _ => return false,
         }
@@ -244,6 +248,7 @@ fn impl_derive_rustsbi_static(name: &Ident, imp: StaticImpl, generics: &Generics
     let cppc_probe: usize = if imp.cppc.is_some() { 1 } else { 0 };
     let nacl_probe: usize = if imp.nacl.is_some() { 1 } else { 0 };
     let sta_probe: usize = if imp.sta.is_some() { 1 } else { 0 };
+    let mpxy_probe: usize = if imp.mpxy.is_some() { 1 } else { 0 };
     let probe = quote! {
         ::rustsbi::_StandardExtensionProbe {
             base: #base_probe,
@@ -258,6 +263,7 @@ fn impl_derive_rustsbi_static(name: &Ident, imp: StaticImpl, generics: &Generics
             cppc: #cppc_probe,
             nacl: #nacl_probe,
             sta: #sta_probe,
+            mpxy: #mpxy_probe,
         }
     };
     let mut match_arms = quote! {};
@@ -335,6 +341,11 @@ fn impl_derive_rustsbi_static(name: &Ident, imp: StaticImpl, generics: &Generics
     if let Some(sta) = &imp.sta {
         match_arms.extend(quote! {
             ::rustsbi::spec::sta::EID_STA => ::rustsbi::_rustsbi_sta(&self.#sta, param, function),
+        })
+    }
+    if let Some(mpxy) = &imp.mpxy {
+        match_arms.extend(quote! {
+            ::rustsbi::spec::mpxy::EID_MPXY => ::rustsbi::_rustsbi_mpxy(&self.#mpxy, param, function),
         })
     }
     let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
@@ -518,6 +529,21 @@ fn impl_derive_rustsbi_dynamic(name: &Ident, imp: DynamicImpl, generics: &Generi
             }
         });
     }
+    let mut mpxy_contents = quote! {};
+    let mut prober_mpxy = quote! {};
+    for mpxy in &imp.mpxy {
+        mpxy_contents.extend(quote! {
+            if ::rustsbi::_rustsbi_mpxy_probe(&self.#mpxy) != ::rustsbi::spec::base::UNAVAILABLE_EXTENSION {
+                return ::rustsbi::_rustsbi_mpxy(&self.#mpxy, param, function)
+            }
+        });
+        prober_mpxy.extend(quote! {
+            let value = ::rustsbi::_rustsbi_mpxy_probe(&self.0.#mpxy);
+            if value != ::rustsbi::spec::base::UNAVAILABLE_EXTENSION {
+                return value
+            }
+        });
+    }
 
     let (_, origin_ty_generics, _) = generics.split_for_impl();
     let prober_generics = {
@@ -546,7 +572,8 @@ fn impl_derive_rustsbi_dynamic(name: &Ident, imp: DynamicImpl, generics: &Generi
                     ::rustsbi::spec::susp::EID_SUSP => { #prober_susp ::rustsbi::spec::base::UNAVAILABLE_EXTENSION },
                     ::rustsbi::spec::cppc::EID_CPPC => { #prober_cppc ::rustsbi::spec::base::UNAVAILABLE_EXTENSION },
                     ::rustsbi::spec::nacl::EID_NACL => { #prober_nacl ::rustsbi::spec::base::UNAVAILABLE_EXTENSION },
-                    ::rustsbi::spec::sta::EID_STA => { #prober_sta ::rustsbi::spec::base::UNAVAILABLE_EXTENSION}
+                    ::rustsbi::spec::sta::EID_STA => { #prober_sta ::rustsbi::spec::base::UNAVAILABLE_EXTENSION},
+                    ::rustsbi::spec::mpxy::EID_MPXY => { #prober_mpxy ::rustsbi::spec::base::UNAVAILABLE_EXTENSION},
                     _ => ::rustsbi::spec::base::UNAVAILABLE_EXTENSION,
                 }
             }
@@ -589,6 +616,7 @@ fn impl_derive_rustsbi_dynamic(name: &Ident, imp: DynamicImpl, generics: &Generi
                     ::rustsbi::spec::cppc::EID_CPPC => { #cppc_contents ::rustsbi::SbiRet::not_supported() },
                     ::rustsbi::spec::nacl::EID_NACL => { #nacl_contents ::rustsbi::SbiRet::not_supported() },
                     ::rustsbi::spec::sta::EID_STA => { #sta_contents ::rustsbi::SbiRet::not_supported() },
+                    ::rustsbi::spec::mpxy::EID_MPXY => { #mpxy_contents ::rustsbi::SbiRet::not_supported() },
                     ::rustsbi::spec::base::EID_BASE => {
                         #define_prober
                         let prober = _Prober(&self);